chore(fixtures): move implmentation into src folder
